pyproject.toml
[project]
name = "qloo_confidential_mcp_server"
version = "0.1.0"
dependencies = [
    "anyio>=4.5",
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "pydantic>=2.7.2,<3.0.0",
    "starlette>=0.27",
    "sse-starlette>=1.6.1",
    "pydantic-settings>=2.5.2",
    "uvicorn>=0.23.1",
    "mcp>=1.11.0",
    "gramine-ratls>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "docker>=7.1.0",
    "jinja2>=3.1.6",
    "tomli>=2.2.1",
    "tomli-w>=1.2.0",
    "pyyaml>=6.0.2",
]